+++ main/decompNormalize.cpp
@@ -410,7 +410,7 @@ void decompressSequence(std::string sequence)
             // End timer after switch-case and print duration
             auto end = std::chrono::steady_clock::now();
             auto elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
-            cout << "Decompression time: " << elapsed / 1000 << " s" << endl;
+            cout << "Decompression time: " << (double)elapsed / 1000 << " s" << endl;
 
             ////////////////////////////////////////////////
             /////////// CPU AND MEM USAGE //////////////////
